Move web3 and proxyProvider to singleton
