+++ src/class/singleton/index.js
@@ -1,4 +1,5 @@
 export { default as httpIdentity } from './httpIdentity';
 export { default as http } from './http';
+export { default as proxyRequest } from './proxyRequest';
 export { default as dappBridge } from './dappBridge';
 export { default as ajv } from './ajv';
