+++ src/class/singleton/proxyRequest/ProxyRequest.js
