+++ src/class/singleton/proxyRequest/decorator/Decorator.js
