+++ src/class/singleton/proxyRequest/decorator/ProviderUrlDecorator.js
