+++ src/class/singleton/proxyRequest/decorator/index.js
