+++ src/class/singleton/proxyRequest/index.js
