+++ src/class/singleton/proxyRequest/provider/ServerProvider.js
