+++ src/class/singleton/proxyRequest/provider/index.js
