+++ src/class/singleton/web3.js
