+++ src/components/pages/Send/TransactionAmountOptions.vue
@@ -95,7 +95,7 @@
 <script>
 import { get } from 'lodash';
 import { BigNumber } from 'bignumber.js';
-import web3 from '@/utils/web3';
+import web3 from '@/class/singleton/web3';
 import VInput from '@/components/ui/form/VInput';
 import VSelect from '@/components/ui/form/VSelect';
