+++ src/services/identityMode.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { proxyRequest } from '@/utils';
+import { proxyRequest } from '@/class/singleton';
 import { NotificationError } from '@/class';
 import { IDENTITY_MODE } from '@/constants';
