+++ src/store/connection-status/getters.js
@@ -1,4 +1,4 @@
-import web3 from '@/utils/web3';
+import web3 from '@/class/singleton/web3';
 
 const appStatus = state => {
   if (!state.web3Connection) {
