+++ src/store/web3/mutations.js
@@ -1,4 +1,4 @@
-import web3 from '@/utils/web3';
+import web3 from '@/class/singleton/web3';
 
 import {
   CHANGE_NETWORK,
