+++ test/unit/specs/store/accounts/actions.spec.js
@@ -1,4 +1,4 @@
-import web3 from '@/utils/web3';
+import web3 from '@/class/singleton/web3';
 import {
   email,
   addresses,
