+++ test/unit/specs/store/transactions/actions.spec.js
@@ -1,4 +1,4 @@
-import web3 from '@/utils/web3';
+import web3 from '@/class/singleton/web3';
 import state from '@/store/transactions';
 import {
   ADD_TRANSACTION,
